+++ app/ui/tab_main.R
@@ -1,9 +1,7 @@
-# QS trend analysis - sidebar
+# QS trend analysis - ui sidebar
 
 tabPanel("Introduction", value = "main",
          htmlOutput("main_text_1"),
          
-         htmlOutput("db_path"),
-         
-         htmlOutput("main_text_2")
+         htmlOutput("paths")
 )